+++ event-subscription/src/test/java/uk/gov/justice/subscription/SubscriptionSorterTest.java
@@ -19,23 +19,23 @@ import org.junit.runner.RunWith;
 import org.mockito.runners.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.class)
-public class SubscriptionHelperTest {
+public class SubscriptionSorterTest {
 
     @Test
     public void shouldSortBySubscriptionPrioritisation() {
 
-        final List<Event> eventList = asList(mock(Event.class),mock(Event.class));
+        final List<Event> eventList = asList(mock(Event.class), mock(Event.class));
 
-        final Subscription subscription1 = new Subscription("name1",eventList,"eventSourceName1", "1");
-        final Subscription subscription2 = new Subscription("name2",eventList,"eventSourceName3", "2");
-        final Subscription subscription3 = new Subscription("name3",eventList,"eventSourceName1", "3");
+        final Subscription subscription1 = new Subscription("name1", eventList, "eventSourceName1", "1");
+        final Subscription subscription2 = new Subscription("name2", eventList, "eventSourceName3", "2");
+        final Subscription subscription3 = new Subscription("name3", eventList, "eventSourceName1", "3");
 
-        final List<Subscription> subscriptions = asList(subscription2,subscription3,subscription1);
+        final List<Subscription> subscriptions = asList(subscription2, subscription3, subscription1);
         final SubscriptionsDescriptor subscriptionsDescriptor = new SubscriptionsDescriptor("specVersion", "service", "serviceComponent", subscriptions);
 
-        final SubscriptionHelper subscriptionHelper = new SubscriptionHelper();
+        final SubscriptionSorter subscriptionSorter = new SubscriptionSorter();
 
-        subscriptionHelper.sortSubscriptionsByPrioritisation(subscriptionsDescriptor);
+        subscriptionSorter.sortSubscriptionsByPrioritisation(subscriptionsDescriptor);
 
         assertThat(subscriptionsDescriptor.getSubscriptions().get(0), is(subscription1));
         assertThat(subscriptionsDescriptor.getSubscriptions().get(1), is(subscription2));
@@ -46,19 +46,19 @@ public class SubscriptionHelperTest {
     @Test
     public void shouldSortBySubscriptionPrioritisationNullsFirst() {
 
-        final List<Event> eventList = asList(mock(Event.class),mock(Event.class));
+        final List<Event> eventList = asList(mock(Event.class), mock(Event.class));
 
-        final Subscription subscription1 = new Subscription("name1",eventList,"eventSourceName1", "1");
-        final Subscription subscription2 = new Subscription("name2",eventList,"eventSourceName3", null);
-        final Subscription subscription3 = new Subscription("name3",eventList,"eventSourceName1", "3");
+        final Subscription subscription1 = new Subscription("name1", eventList, "eventSourceName1", "1");
+        final Subscription subscription2 = new Subscription("name2", eventList, "eventSourceName3", null);
+        final Subscription subscription3 = new Subscription("name3", eventList, "eventSourceName1", "3");
 
-        final List<Subscription> subscriptions = asList(subscription2,subscription3,subscription1);
+        final List<Subscription> subscriptions = asList(subscription2, subscription3, subscription1);
 
         final SubscriptionsDescriptor subscriptionsDescriptor = new SubscriptionsDescriptor("specVersion", "service", "serviceComponent", subscriptions);
 
-        final SubscriptionHelper subscriptionHelper = new SubscriptionHelper();
+        final SubscriptionSorter subscriptionSorter = new SubscriptionSorter();
 
-        subscriptionHelper.sortSubscriptionsByPrioritisation(subscriptionsDescriptor);
+        subscriptionSorter.sortSubscriptionsByPrioritisation(subscriptionsDescriptor);
 
         assertThat(subscriptionsDescriptor.getSubscriptions().get(0), is(subscription2));
         assertThat(subscriptionsDescriptor.getSubscriptions().get(1), is(subscription1));
@@ -69,22 +69,22 @@ public class SubscriptionHelperTest {
     @Test
     public void shouldThrowExceptionIfPrioritisationNotNumber() {
 
-        final List<Event> eventList = asList(mock(Event.class),mock(Event.class));
+        final List<Event> eventList = asList(mock(Event.class), mock(Event.class));
 
-        final Subscription subscription1 = new Subscription("name1",eventList,"eventSourceName1", "1");
-        final Subscription subscription2 = new Subscription("name2",eventList,"eventSourceName3", "wrongPrioritisation");
-        final Subscription subscription3 = new Subscription("name3",eventList,"eventSourceName1", "3");
+        final Subscription subscription1 = new Subscription("name1", eventList, "eventSourceName1", "1");
+        final Subscription subscription2 = new Subscription("name2", eventList, "eventSourceName3", "wrongPrioritisation");
+        final Subscription subscription3 = new Subscription("name3", eventList, "eventSourceName1", "3");
 
-        final List<Subscription> subscriptions = asList(subscription2,subscription3,subscription1);
+        final List<Subscription> subscriptions = asList(subscription2, subscription3, subscription1);
 
         final SubscriptionsDescriptor subscriptionsDescriptor = new SubscriptionsDescriptor("specVersion", "service", "serviceComponent", subscriptions);
 
-        final SubscriptionHelper subscriptionHelper = new SubscriptionHelper();
+        final SubscriptionSorter subscriptionSorter = new SubscriptionSorter();
 
         try {
-            subscriptionHelper.sortSubscriptionsByPrioritisation(subscriptionsDescriptor);
+            subscriptionSorter.sortSubscriptionsByPrioritisation(subscriptionsDescriptor);
             fail();
-        }catch (final YamlParserException expected){
+        } catch (final YamlParserException expected) {
             assertThat(expected.getMessage(), is("Incorrect prioritisation number: wrongPrioritisation in subscription-descriptor.yaml"));
         }
     }