+++ event-subscription/src/test/java/uk/gov/justice/subscription/domain/eventsource/DefaultEventSourceDefinitionFactoryTest.java
@@ -1,5 +1,6 @@
 package uk.gov.justice.subscription.domain.eventsource;
 
+import static java.util.Optional.empty;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.fail;
@@ -34,13 +35,13 @@ public class DefaultEventSourceDefinitionFactoryTest {
 
         final Location location = eventSourceDefinition.getLocation();
 
-        if(location.getDataSource().isPresent()) {
+        if (location.getDataSource().isPresent()) {
             assertThat(location.getDataSource().get(), is("java:/app/my-context-service/DS.eventstore"));
         } else {
             fail();
         }
 
         assertThat(location.getJmsUri(), is("JMS URI not used"));
-        assertThat(location.getRestUri(), is("Rest URI not used"));
+        assertThat(location.getRestUri(), is(empty()));
     }
 }