+++ CHANGELOG.md
@@ -4,6 +4,8 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- rest_uri in event-sources.yaml is now optional to support JMS only sources
 
 ## [6.0.0-RC2] - 2019-07-04
 ### Added